transpile: whole-function inference, the durable mismatch and the kinded literal; a contradiction poisons what it touched, an integer literal has a kind, an undecided candidate never beats a decided one
Eight items of 2026-09-08-addenda-inference-3.md's FINAL ORDER for the fourth
agent (report 2026-09-08-report-inference-4.md, progress log
2026-09-08-progress-inference-4.md), over 3a2b76a: the findings of the two
reviews of 9099480 and the two of 8ab991f that the third agent did not take.

A constraint with no solution is a fact the table keeps, re-checked at the
fixed point, poisoning every variable it touched and what it was read
through, and reported ONCE from the table after the quiet pre-pass's rewind;
a poisoned variable answers as an unknown, so the local takes the untyped
path instead of keeping the first binding. Two solved types are compared
(the raw-variable gate is gone), the projection suppression is narrowed to
the component actually unread, and only the solve reports: the writing walk
binds nothing and its own reporting is retired, since over the corpus it
filed 32 rows of the engine's own reading and none the solve had not seen.
An unsuffixed integer or float literal mints a variable with an integral or
float KIND that refuses a binding the kind does not admit and defaults to
i32 or f64 only after the fixed point, which is Rust's own rule: `let mut n
= 0; n = xs.len()` is usize and the i32 overflow the review showed is gone;
comparing solved types could not be honest before it, because 310 of 464
corpus contradictions were the engine's own default meeting a real width.
A type mentioning a variable is not settled, so a closure's body supplies
its result to the callable bound. A bound projects through the type as
written, never the peeled one. A conditional candidate never beats a decided
one at any deref depth. An await answers only where its output can be read,
and one that did not type consumes nothing. The occurs check reads through
the table. The pre-pass binds a body's own consts. A callable position sees a
closure through the block that prepares its captures.

Contract choices 4.1-4.6 of the report accepted: only the solve reports a
contradiction (retiring the third commit's scratch-copy reporting); a value
and what it derefs to do not disagree, at a coercion site; a type still
carrying an unknown at the fixed point is evidence neither way; a
contradiction poisons what it was read through; the solved-type comparison
lands with the kinded literal; a closure behind its capture block stands in
a callable position. Verified closed: OO4 (the where-clause tie is an
ambiguity) and OO8's struct-field escape. Five of the six corpus sites the
review named as erased contradictions do not reproduce at 3a2b76a; the sixth
(websocket-client-wasm client.rs:135) reports and is then withdrawn by the
deref rule, correctly.

Found and recorded, not fixed: a refused expression answers the unit type,
which is what makes any new constraint source dangerous (two attempts to
read a callable position's result from inside the closure's scope filed five
false contradictions and were taken out) and what emits two `(value: void)`
parameters in a signals test file; the ambiguity message cannot tell two
impls apart that differ only in their where clause; a method that moves a
field out of self and then drops self double-drops it; a conditional impl's
bound is put on the class.

Budget: ankql 136 -> 133, signals 115 -> 112, core 1556 -> 1576 (+22
byte-buffer holes the now-correct u8 element exposes, -5 index rows, +4
durable mismatches, +2 await refusals, -3 obligations); proto 62. Two goldens
with drivers failing against 3a2b76a (a fatal double drop, a wrong value);
nine golden headers cut of their commit history; six files that crossed 600
lines split. Harness 997 -> 1019 across fifteen targets, zero warnings; proto
301/1, ankql 422/0 on the tree, base 535/0, zero leak reports and zero fatal
markers; the ten-crate emission delta five files, eight hunks, each
attributed. Verified in the quiet worktree and on main after the merge.

+++ transpile/goldens/a_cursors_two_axes/expected.ts
@@ -90,7 +90,7 @@ export function anyThenCount<I extends Iterable<Token>>(walk: SeqCursor<Token>):
 export function countRefs<I extends Iterable<Token>>(walk: SeqCursor<Token>): number {
   let n = 0;
   for (const _t of walk.takeRest()) {
-    n = checkedAdd(n, 1, 'i32');
+    n = checkedAdd(n, 1, 'usize');
   }
   return n;
 }
